requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Background processing
redis>=5.0.0
celery>=5.3.0

# Cloud storage
boto3>=1.34.0

# Logging
structlog>=23.2.0

# File handling
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.0

# 3D mesh processing
trimesh>=4.0.0
numpy>=1.24.0
scipy>=1.11.0

# HTTP client
httpx>=0.25.0

# Environment management
python-dotenv>=1.0.0
